refactor: split ooda_actions.rs (1,399 LOC) into module directory
Split the oversized ooda_actions.rs into src/ooda_actions/ with 7 sub-modules:
- mod.rs: dispatch entry points, helpers (make_outcome, next_progress)
- simple_actions.rs: consolidate, research, improve, gym, skill dispatchers
- advance_goal.rs: goal routing, subordinate heartbeat checking
- goal_session.rs: session-based goal advancement with base-type agent
- verification.rs: action verification, progress parsing
- session.rs: PTY terminal session launcher
- test_helpers.rs: shared mock infrastructure (#[cfg(test)])

All 28 tests pass. No public API changes.
